requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
mysql-connector-python
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]
passlib[bcrypt]
python-multipart
fastapi-mail